backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
openai
python-dotenv
sqlalchemy
pyarrow
alembic
pytest
httpx
